Make the default inventories named inventory.ini
So I like the fact that the default inventory for the demo was not named
just 'hosts' because that is confusing in normal text (easily overlooked
that it is a file), but also because it doesn't just include hosts, also
groups and variables.

So the thing I would definitely change is adding the extension so it
stands out even more as a file, but also the content is clear. And you
may get (some) syntax highlighting out of it, although it's not an
actual INI file format though...